requirements.txt
fastembed
fastapi
uvicorn
python-multipart
jinja2
PyPDF2
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-core
faiss-cpu
python-dotenv
